# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Chat Message Management
  version: 1.0.0
  description: |-
    ## Basic Chat Message API
    REST based service for managing chat messaging
    ### Resource
    - message
    ### Operations
    - Create new entity
    - Retrieve the whole set of entities
tags:
  - name: ChatMessage
paths:
  /messages:
    get:
      operationId: listMesasges
      summary: 'Lists the whole list of chat messages.'
      description: 'This operation retrieves the list of all messages.'
      tags:
        - ChatMessage
      responses:
        '200':
          description: 'Success'
          content:
            application/json;charset=utf-8:
              schema:
                items:
                  $ref: '#/components/schemas/Message'
                type: array
        '500':
          description: 'Internal Server Error'
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createMesasge
      summary: 'Creates a new chat message.'
      description: 'This operation a new chat message.'
      tags:
        - ChatMessage
      requestBody:
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/Message'
        description: 'Content of the chat message to be created'
        required: true
      responses:
        '201':
          description: 'Created'
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: 'Bad Request'
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 'Internal Server Error'
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Message:
      type: object
      properties:
        id:
          description: 'Unique identifier of the entity. Generated automatically.'
          type: string
          example: '123e4567-e89b-12d3-a456-426655440000'
          readOnly: true
        author:
          description: 'Name of the person, who posted the message.'
          type: string
          example: 'John Dow'
        text:
          description: 'Text of the message.'
          type: string
          example: 'Yesterday all my troubles seemed so far away'
        date:
          description: 'Date, when the message was posted. Generated automatically.'
          type: string
          format: date-time
          example: '2022-09-23T18:17:28Z'
          readOnly: true
        '@type':
          description: 'Type of the entity.'
          type: string
          enum:
            - Message
      required:
        - id
        - author
        - text
        - date
        - '@type'
    Error:
      type: object
      properties:
        status:
          description: 'The HTTP status code generated by the origin server for this occurrence of the problem. Generators MUST use the same status code in the actual HTTP response.'
          type: integer
        title:
          description: 'A short, human-readable summary of the problem type.'
          type: string
        details:
          description: 'A human-readable explanation specific to this occurrence of the problem.'
          type: string
      required:
        - status
        - title